requirements-dev.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements-dev.in
#
aiohappyeyeballs==2.4.8
    # via aiohttp
aiohttp==3.11.13
    # via elasticsearch
aiosignal==1.3.2
    # via aiohttp
alembic==1.12.1
    # via -r D:\crown-nexus\requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   httpx
    #   starlette
    #   watchfiles
asyncpg==0.28.0
    # via -r D:\crown-nexus\requirements.in
attrs==25.1.0
    # via
    #   aiohttp
    #   hypothesis
bcrypt==4.3.0
    # via passlib
black==23.9.1
    # via -r requirements-dev.in
build==1.2.2.post1
    # via pip-tools
certifi==2025.1.31
    # via
    #   elastic-transport
    #   httpcore
    #   httpx
cffi==1.17.1
    # via cryptography
cfgv==3.4.0
    # via pre-commit
click==8.1.8
    # via
    #   black
    #   pip-tools
    #   uvicorn
colorama==0.4.6
    # via
    #   build
    #   click
    #   pytest
    #   uvicorn
coverage[toml]==7.6.12
    # via pytest-cov
cryptography==44.0.2
    # via
    #   python-jose
    #   types-pyopenssl
    #   types-redis
distlib==0.3.9
    # via virtualenv
dnspython==2.7.0
    # via email-validator
ecdsa==0.19.0
    # via python-jose
elastic-transport==8.17.0
    # via elasticsearch
elasticsearch[async]==8.10.1
    # via -r D:\crown-nexus\requirements.in
email-validator==2.2.0
    # via pydantic
fastapi==0.104.1
    # via -r D:\crown-nexus\requirements.in
filelock==3.17.0
    # via virtualenv
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.1.1
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.25.2
    # via
    #   -r D:\crown-nexus\requirements.in
    #   -r requirements-dev.in
hypothesis==6.86.2
    # via -r requirements-dev.in
identify==2.6.8
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   yarl
iniconfig==2.0.0
    # via pytest
isort==5.12.0
    # via -r requirements-dev.in
mako==1.3.9
    # via alembic
markupsafe==3.0.2
    # via mako
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy==1.5.1
    # via -r requirements-dev.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
nodeenv==1.9.1
    # via pre-commit
packaging==24.2
    # via
    #   black
    #   build
    #   pytest
passlib[bcrypt]==1.7.4
    # via -r D:\crown-nexus\requirements.in
pathspec==0.12.1
    # via black
pip-tools==7.3.0
    # via -r requirements-dev.in
platformdirs==4.3.6
    # via
    #   black
    #   virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==3.4.0
    # via -r requirements-dev.in
propcache==0.3.0
    # via
    #   aiohttp
    #   yarl
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycparser==2.22
    # via cffi
pydantic[email]==2.4.2
    # via
    #   -r D:\crown-nexus\requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.10.1
    # via pydantic
pydantic-settings==2.0.3
    # via -r D:\crown-nexus\requirements.in
pyproject-hooks==1.2.0
    # via build
pytest==7.4.4
    # via
    #   -r requirements-dev.in
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==0.21.2
    # via -r requirements-dev.in
pytest-cov==4.1.0
    # via -r requirements-dev.in
python-dotenv==1.0.1
    # via
    #   pydantic-settings
    #   uvicorn
python-jose[cryptography]==3.3.0
    # via -r D:\crown-nexus\requirements.in
python-multipart==0.0.20
    # via -r D:\crown-nexus\requirements.in
pyyaml==6.0.2
    # via
    #   pre-commit
    #   uvicorn
redis==5.0.8
    # via -r D:\crown-nexus\requirements.in
rsa==4.9
    # via python-jose
ruff==0.0.292
    # via -r requirements-dev.in
six==1.17.0
    # via ecdsa
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sortedcontainers==2.4.0
    # via hypothesis
sqlalchemy[asyncio]==2.0.38
    # via
    #   -r D:\crown-nexus\requirements.in
    #   alembic
starlette==0.27.0
    # via fastapi
tenacity==8.2.3
    # via -r D:\crown-nexus\requirements.in
types-cffi==1.16.0.20241221
    # via types-pyopenssl
types-passlib==1.7.7.20241221
    # via -r D:\crown-nexus\requirements.in
types-pyasn1==0.6.0.20250208
    # via types-python-jose
types-pyopenssl==24.1.0.20240722
    # via types-redis
types-python-jose==3.3.4.20240106
    # via -r D:\crown-nexus\requirements.in
types-redis==4.6.0.20241004
    # via -r D:\crown-nexus\requirements.in
types-setuptools==75.8.2.20250305
    # via types-cffi
typing-extensions==4.12.2
    # via
    #   alembic
    #   fastapi
    #   mypy
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
urllib3==2.3.0
    # via elastic-transport
uvicorn[standard]==0.23.2
    # via -r D:\crown-nexus\requirements.in
virtualenv==20.29.2
    # via pre-commit
watchfiles==1.0.4
    # via uvicorn
websockets==15.0.1
    # via uvicorn
wheel==0.45.1
    # via pip-tools
yarl==1.18.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
